# web/openapi.json
{
  "components": {
    "schemas": {
      "Body_createCloudTask": {
        "properties": {
          "config": {
            "description": "任务配置，JSON 对象",
            "title": "Config",
            "type": "string"
          },
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "tool_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Tool Id",
            "type": "string"
          }
        },
        "required": [
          "tool_id",
          "name",
          "config",
          "files"
        ],
        "title": "Body_createCloudTask",
        "type": "object"
      },
      "CloudFileOut": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "size",
          "content_type"
        ],
        "title": "CloudFileOut",
        "type": "object"
      },
      "CloudTaskDetail": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CloudFileOut"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "string"
          },
          "total_bytes": {
            "title": "Total Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tool_id",
          "name",
          "total_bytes",
          "created_at",
          "files",
          "config"
        ],
        "title": "CloudTaskDetail",
        "type": "object"
      },
      "CloudTaskSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CloudFileOut"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "string"
          },
          "total_bytes": {
            "title": "Total Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tool_id",
          "name",
          "total_bytes",
          "created_at",
          "files"
        ],
        "title": "CloudTaskSummary",
        "type": "object"
      },
      "CodeLoginRequest": {
        "properties": {
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email",
          "code"
        ],
        "title": "CodeLoginRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Health": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "title": "Health",
        "type": "object"
      },
      "Me": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "has_password": {
            "title": "Has Password",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "has_password",
          "created_at"
        ],
        "title": "Me",
        "type": "object"
      },
      "PasswordLoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "PasswordLoginRequest",
        "type": "object"
      },
      "Readiness": {
        "properties": {
          "database": {
            "title": "Database",
            "type": "boolean"
          },
          "storage": {
            "title": "Storage",
            "type": "boolean"
          }
        },
        "required": [
          "database",
          "storage"
        ],
        "title": "Readiness",
        "type": "object"
      },
      "SendCodeRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "SendCodeRequest",
        "type": "object"
      },
      "SetPasswordRequest": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "SetPasswordRequest",
        "type": "object"
      },
      "Usage": {
        "properties": {
          "max_file_bytes": {
            "title": "Max File Bytes",
            "type": "integer"
          },
          "max_files_per_task": {
            "title": "Max Files Per Task",
            "type": "integer"
          },
          "max_tasks": {
            "title": "Max Tasks",
            "type": "integer"
          },
          "quota_bytes": {
            "title": "Quota Bytes",
            "type": "integer"
          },
          "task_count": {
            "title": "Task Count",
            "type": "integer"
          },
          "used_bytes": {
            "title": "Used Bytes",
            "type": "integer"
          }
        },
        "required": [
          "used_bytes",
          "quota_bytes",
          "task_count",
          "max_tasks",
          "max_file_bytes",
          "max_files_per_task"
        ],
        "title": "Usage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Make Your Life Better API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/account": {
      "delete": {
        "description": "删除账号，连同所有云端文件。",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/code": {
      "post": {
        "description": "发送登录验证码。邮箱还没注册时，用验证码登录就会自动注册。",
        "operationId": "sendLoginCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "发送太频繁"
          },
          "503": {
            "description": "邮件发送失败"
          }
        },
        "summary": "Send Code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/code": {
      "post": {
        "operationId": "loginWithCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "验证码错误或过期"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login With Code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/password": {
      "post": {
        "operationId": "loginWithPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "邮箱或密码不正确"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "暂时锁定"
          }
        },
        "summary": "Login With Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "getMe",
        "parameters": [
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "未登录"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password": {
      "put": {
        "description": "设置或修改密码（已经登录就能改；忘记密码时先用验证码登录）。其他设备会被登出。",
        "operationId": "setPassword",
        "parameters": [
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/cloud/tasks": {
      "get": {
        "operationId": "listCloudTasks",
        "parameters": [
          {
            "in": "query",
            "name": "tool_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tool Id"
            }
          },
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CloudTaskSummary"
                  },
                  "title": "Response Listcloudtasks",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "cloud"
        ]
      },
      "post": {
        "operationId": "createCloudTask",
        "parameters": [
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createCloudTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudTaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "超出额度"
          },
          "413": {
            "description": "文件太大"
          },
          "422": {
            "description": "参数不对"
          }
        },
        "summary": "Create Task",
        "tags": [
          "cloud"
        ]
      }
    },
    "/api/cloud/tasks/{task_id}": {
      "delete": {
        "operationId": "deleteCloudTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Task",
        "tags": [
          "cloud"
        ]
      },
      "get": {
        "operationId": "getCloudTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudTaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "cloud"
        ]
      }
    },
    "/api/cloud/tasks/{task_id}/files/{file_id}": {
      "get": {
        "operationId": "downloadCloudFile",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "File Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download File",
        "tags": [
          "cloud"
        ]
      }
    },
    "/api/cloud/usage": {
      "get": {
        "operationId": "getCloudUsage",
        "parameters": [
          {
            "in": "cookie",
            "name": "mylb_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mylb Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage",
        "tags": [
          "cloud"
        ]
      }
    },
    "/api/health": {
      "get": {
        "description": "存活检查：进程在就返回 ok，不检查依赖。",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/health/ready": {
      "get": {
        "description": "就绪检查：数据库和对象存储都能访问才返回 200。",
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    }
  }
}